third_party/blink/web_tests/fast/css/invalidation/required-optional-pseudo-expected.txt

Use descendant invalidation sets for :required and :optional pseudo classes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(inputOptional, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(inputOptional, '').backgroundColor is green
PASS getComputedStyle(inputRequired, '').backgroundColor is transparent
PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS getComputedStyle(inputRequired, '').backgroundColor is green
PASS successfullyParsed is true

TEST COMPLETE
